Move the media and transcription routes into serve/openai.rs
